Have the shader-setting routines (Shader, ConnectShaders, ShaderGroupBegin, ShaderGroupEnd, Parameter) all return a bool for error status, and beef up a few of their error messages.
Also, fix important bug wherein Shader was not checking to see if the
current attrib state was already being used before writing to it.  If
somebody else (like a piece of geometry) is holding a reference to it,
then we need to create a new attrib state to modify for whomever comes
along next to want it.  Not doing this properly prevents us from having
more than one shader in the scene.


git-svn-id: https://openshadinglanguage.googlecode.com/svn/trunk@298 a3683778-fedd-11de-87c3-653bb6514e15