Significant additional work on "runtime code optimization" introduced recently. Changes include:
* New constant folding for additional ops: neg, aref (was previously
commented out due to bugs), compref, strlen, endswith, concat, clamp,
sqrt, matrix, transform[nv].

* All previous optimizations had been ones where no *new* constant
values were needed (we could do new constant symbols, but only using
VALUES that were stored statically or were the defaults of parameters
stored in the master).  Now we can make up brand new constants with new
values, which are stored in the ShadingSystem via the helper class
ConstantPool.  This allows add, sub, mul, and div where both operands
are constant will just turn into a single constant, as well as several
of the other new constant folding routines to work.

* Refactor to eliminate the if-else-else-else of constant folding
routine selection, in favor of a hash map acting as a dispatch table.

* Allow "local aliasing" within basic blocks -- previously, "A=const"
would remember that A had the constant value (so subsequent C=A*B, say,
would treat A as a constant) but only if the statement was in the global
scope and A was never subsequently reassigned.  Now we can remember that
A has that value, but just for the basic block it's in and/or until it's
reassigned.  This allows lots of additional constant folding.  When each
basic block is done, its local aliases are forgotten.

* Some preliminary groundwork for tracking which params will be used by
downstream layers, so I can take advantage of that layer.  (Apologies
for the incomplete nature of that.)

* Add -O0, -O1, -O2 optional flags to "testshade" to specify level of
runtime optimization.

* New statistics track how many actual instructions we execute over a
render, and how much total time we spend doing the runtime optimization.

* A couple minor renames and cleanups I promised Chris I'd do in the
last review, but somehow botched the checkin and didn't get them all
committed.

* IMPORTANT bug fix to the last checkin, in constfold_add -- lines 340
and 347 got the argument numbers mixed up, resulting in turning "A = B +
0" into "A=0" rather than "A=B".  Oops!

So, the net result of all this is summarized by the following stats from
a production asset we like to use as a benchmark:

optimize=1  5:35  2046 MB peak  1,806,370,431 instructions run
optimize=2  4:32  1403 MB peak  1,046,991,997 instructions run

As for how we can reduce the number of instructions executed by almost
45% but only reduce runtime by 20%, let's just say there are other
bottlenecks in our renderer we're working on and interpreter overhead is
obviously not the major (or at least not the sole) bottleneck.

http://codereview.appspot.com/639042/show
Review by: ckulla


git-svn-id: https://openshadinglanguage.googlecode.com/svn/trunk@605 a3683778-fedd-11de-87c3-653bb6514e15