Improve OSL memory usage on several fronts.
(0) Instrumented with lots of new memory stats to track where all the
shading memory is going.

(1) We were retaining HUGE (pre-optimized) symbol and code vectors on
each instance.  This made a big peak memory footprint, despite freeing
them as we optimize the groups (since the peak is hit right away).  So
I've modified the instance to only copy only the param part of the
symbol vector, and not the ops and args at all, when constructing from
the master.  The rest of the symbols, and ops/args, are copied when it's
time to optimize that group; and after optimization they are, as before,
trimmed down.  This greatly reduced the peak memory that instances
consume to about 10% of its previous peak.

(2) Fixed the longstanding problem of endlessly accumulating cruft into
the same llvm Module and ExecutionEngine in the following way: (a)
Create a DefaultJITMemoryManager that we retain for the whole render.
(b) Have a DummyJITMemoryManager that passes all the calls through to
JITMM, except for the deallocations, which it drops on the floor.  Also,
when it destroys, it doesn't destroy the permanent JMM. (c) When we
generate IR and JIT for a shader group, we create a totally new Module,
ExecutionEngine, and DummyMM every time.  After the JIT is done and we
extract the function pointer (remember the code lives in the retained
MM), we free all of these.

Also, I totally ripped out the old ShadingExecution after carefully
removing all remaining references to it.

Review by: ckulla



git-svn-id: https://openshadinglanguage.googlecode.com/svn/trunk@869 a3683778-fedd-11de-87c3-653bb6514e15