Optimize derivs zeroing for pointcloud
Sometimes we have to clear the derivs in the output
arrays of pointcloud_search and always for pointcloud_get.
But there is no need to clear more than the actual count
of elements being returned. This can be a big speedup if
somebody calls these functions with 1k element arrays but
only few results are returned most of the time.