More runtime optimization debugging aids: several new options allow individual optimizations to be disabled (to quickly narrow down potential problems when debugging runtime optimizations).
A new "optimize_nondebug" option, when used in conjunction with debug=1
and a named debug_groupname, will cause all the other (not being
debugged) groups to run with full optimization.  This helps speed up
debug runs when comparing one group optimized and not optimized, if
everything else is also unoptimized, the scenes take too long to render.

Also some very minor shifting around of the check for "only_groupname",
so that the stats correctly reflect the number of shader groups that were
fully compiled.