Layer-to-layer global propagation
We were already good at noticing that a connected parameter's upstreadm
output is guaranteed to be a constant, and then replace the downstream
parameter with that constant (no longer needing the connection at all).

Now we also add logic to notice cases where the upstream output is
guaranteed to be assigned a global variable (such as N or u), and
we can alias the downstream parameter to that global, just as if it
was assigned the global as a default argument in the same shader (also
no longer needing the connection at all).